Migrate ImageButton sample
