+++ samples/ControlGallery/ControlGallery/MainPage.xaml
@@ -59,6 +59,10 @@
 						          Detail="Button control"
 						          Command="{Binding NavigateCommand}"
 						          CommandParameter="{x:Type pages:ButtonPage}"/>
+						<TextCell Text="ImageButton"
+						          Detail="ImageButton control"
+						          Command="{Binding NavigateCommand}"
+						          CommandParameter="{x:Type pages:ImageButtonPage}"/>
 						<TextCell Text="CheckBox"
 								  Detail="CheckBox control for selections"
 								  Command="{Binding NavigateCommand}"
